commercial-search/requirements.txt
# requirements.txt
# Install via `pip3 install -r requirements.txt`
html2text
wikitextparser
flask
flask_sqlalchemy
python-dotenv
click
tqdm
xml-sitemap-writer